Consolidate helper functions/classes in helpers module
