+++ src/tests/unit/nodes/test_betting.py
@@ -8,8 +8,8 @@ from tests.fixtures.data_factories import (
     fake_footywire_betting_data,
     fake_cleaned_match_data,
 )
+from tests.helpers import ColumnAssertionMixin
 from augury.nodes import betting
-from .node_test_mixins import ColumnAssertionMixin
 
 
 N_MATCHES_PER_SEASON = 4
